Ifpack2 Hypre: Fix link errors when multiple node types are used
