+++ packages/ifpack2/src/Ifpack2_Hypre.cpp
@@ -1,0 +1,78 @@
+/*@HEADER
+// ***********************************************************************
+//
+//       Ifpack2:  Templated Object-Oriented Algebraic Preconditioner Package
+//                 Copyright (2009) Sandia Corporation
+//
+// Under terms of Contract DE-AC04-94AL85000, there is a non-exclusive
+// license for use of this work by or on behalf of the U.S. Government.
+//
+// Redistribution and use in source and binary forms, with or without
+// modification, are permitted provided that the following conditions are
+// met:
+//
+// 1. Redistributions of source code must retain the above copyright
+// notice, this list of conditions and the following disclaimer.
+//
+// 2. Redistributions in binary form must reproduce the above copyright
+// notice, this list of conditions and the following disclaimer in the
+// documentation and/or other materials provided with the distribution.
+//
+// 3. Neither the name of the Corporation nor the names of the
+// contributors may be used to endorse or promote products derived from
+// this software without specific prior written permission.
+//
+// THIS SOFTWARE IS PROVIDED BY SANDIA CORPORATION "AS IS" AND ANY
+// EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+// IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR
+// PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL SANDIA CORPORATION OR THE
+// CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
+// EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
+// PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
+// PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF
+// LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
+// NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+// SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+//
+// Questions? Contact Michael A. Heroux (maherou@sandia.gov)
+//
+// ***********************************************************************
+//@HEADER
+*/
+
+#include "Ifpack2_ConfigDefs.hpp"
+#if defined(HAVE_IFPACK2_HYPRE) && defined(HAVE_IFPACK2_MPI)
+
+#include "HYPRE_IJ_mv.h"
+#include "HYPRE_parcsr_ls.h"
+#include "krylov.h"
+#include "_hypre_parcsr_mv.h"
+#include "_hypre_IJ_mv.h"
+#include "HYPRE_parcsr_mv.h"
+#include "HYPRE.h"
+#include <map>
+#include "Ifpack2_Hypre_FunctionParameters.hpp"
+
+namespace Ifpack2 {
+
+  #include "Ifpack2_HypreParameterMap.hpp"
+
+  void IFPACK2_CHK_ERRV(int code) {
+    if(code<0) {
+      std::ostringstream ofs;
+      ofs << "Ifpack2::Hypre: Error with code "<<code<<std::endl;
+      throw std::runtime_error(ofs.str());
+    }
+  }
+
+  void IFPACK2_CHK_ERR(int code) {
+    if(code<0) {
+      std::ostringstream ofs;
+      ofs << "Ifpack2::Hypre: Error with code "<<code<<std::endl;
+      throw std::runtime_error(ofs.str());
+    }
+  }
+
+}
+
+#endif // HAVE_IFPACK2_HYPRE && HAVE_IFPACK2_MPI
